requirements.txt
pyyaml
smbus
